Added wishlist delete functionality
